refactor: make files smaller and more focused
