--- scripts/migrate-v9/components.mjs
@@ -1,70 +1,0 @@
-// Find components who changed to an html element
-export const componentTypes = [
-  'Accordion',
-  'Alert.Button',
-  'Alert.SecondaryButton',
-  'Alert.Title',
-  'Alert',
-  'Box',
-  'AspectRatio',
-  'Avatar',
-  'Badge',
-  'Breadcrumb',
-  'Button',
-  'ButtonGroup',
-  'CloseButton',
-  'Field',
-  'Flex',
-  'Grid',
-  'InputText',
-  'Label',
-  'Link',
-  'Stack',
-  'Text',
-  'Textarea',
-  'Toggle',
-]
-
-export // Find components who did not changed of element but only move props style to className
-const componentTypesUnchanged = [
-  'Accordion',
-  'Alert',
-  'Alert.Button',
-  'Alert.SecondaryButton',
-  'Alert.Title',
-  'AspectRatio',
-  'Avatar',
-  'Badge',
-  'Breadcrumb',
-  'Button',
-  'ButtonGroup',
-  'Card',
-  'Checkbox',
-  'CloseButton',
-  'Drawer',
-  'DropdownMenu',
-  'Field',
-  'FileDrop',
-  'Files',
-  'Form',
-  'Hint',
-  'IconsFont',
-  'InputText',
-  'Label',
-  'Link',
-  'Modal',
-  'Pagination',
-  'Popover',
-  'Radio',
-  'Select',
-  'Swiper',
-  'Table',
-  'Tabs',
-  'Tag',
-  'Text',
-  'Textarea',
-  'Toast',
-  'Toggle',
-  'Tooltip',
-  'UniversalLink',
-]
