+++ scripts/migrate-v9/fixtures/Component2.tsx
@@ -1,4 +1,14 @@
-import { Box } from '../../../lib/src/old/components/Box'
+import { Breadcrumb } from '../../../lib/src/old/components/Breadcrumb'
 
 // Usage examples
-export const Example = () => <Box display="flex" h="100%"></Box>
+export const Example = () => {
+  return (
+    <Breadcrumb mb="lg" separator=">">
+      <Breadcrumb.Item href="/" mb="xl">
+        Introduction
+      </Breadcrumb.Item>
+      <Breadcrumb.Item mb="xl">Disabled</Breadcrumb.Item>
+      <Breadcrumb.Item mb="xl">Breadcrumb</Breadcrumb.Item>
+    </Breadcrumb>
+  )
+}
